--- a/packages/api-gen/bin/api-gen.mjs
+++ b/packages/api-gen/bin/api-gen.mjs
@@ -1,3 +1,2 @@
 #!/usr/bin/env node
-"use strict";
 import "../dist/cli.mjs";