--- a/packages/api-gen/eslint.config.mjs
+++ b/packages/api-gen/eslint.config.mjs
@@ -1,3 +1,0 @@
-import shopwareConfig from "eslint-config-shopware";
-
-export default [...shopwareConfig];