--- a/packages/api-gen/src/processAstSchemaAndOverrides.ts
+++ b/packages/api-gen/src/processAstSchemaAndOverrides.ts
@@ -1,6 +1,6 @@
 import {
-  GenerationMap,
-  TransformedElements,
+  type GenerationMap,
+  type TransformedElements,
   generateFile,
 } from "./generateFile";
 import { createVirtualFiles } from "./virtualFileCreator";