switched `api-gen` package linting to biome for better accuracy and performance
test: time `pnpm run lint` inside package

before
```
pnpm run lint  4.95s user 0.52s system 150% cpu 3.644 total
```

after
```
pnpm run lint  0.33s user 0.13s system 85% cpu 0.540 total
```
apart of the speed it was more accurate in finding problems and making the code more consistent